lab8/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.11.1
asttokens==3.0.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==6.0.3
executing==2.2.0
fastapi==0.121.2
filelock==3.20.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
javascript==1!1.2.6
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
line_profiler==4.2.0
lxml==5.4.0
magic-filter==1.0.12
matplotlib==3.10.1
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-telegram-bot==21.10
pytz==2025.2
pywin32==310
PyYAML==6.0.3
pyzmq==26.4.0
qtconsole==5.6.1
QtPy==2.4.3
regex==2025.11.3
requests==2.32.3
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.49.3
threadpoolctl==3.5.0
tokenizers==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3

django
djangorestframework